backend/requirements.txt
flask == 2.2.3
flask_restx == 1.0.6
Flask_Cors==3.0.10
Flask_JWT_Extended==4.4.4
flask_pydantic==0.11.0
mongoengine == 0.26.0
flask_mongoengine == 1.0.0
pydantic == 1.10.5
pipreqs == 0.4.11
bcrypt==4.0.1
python_json_logger==2.0.7
boto3 == 1.26.127
python-dotenv==0.21.0
pymongo==4.2.0
